requirements.txt
# HTTP client with minimal dependencies - compression disabled by default
boto3==1.38.18
openshift==0.13.2
openstacksdk==4.5.0
httpx==0.28.1
python-dotenv==1.1.0
slack_bolt==1.23.0
pytest==8.3.5
dynaconf[vault]==3.2.11

# SDK dependencies
gspread==6.2.1
requests>=2.28.0
packaging>=21.0
python-dateutil>=2.8.2

# Slack SDK - REQUIRED for ROTA notifications (WebClient, SlackApiError)
# Used by slack_client.py for: chat_postMessage (channels), conversations_open (DMs)
# Without this, group reminders and DM notifications will fail
slack-sdk==3.35.0
